feat: enable dry run mode

+++ main.sh
@@ -1,6 +1,10 @@
 #! /bin/sh
 
-semantic-release -d | tee semantic-release-output.txt
+if [ "$DRY_RUN" != "true" ]; then
+  semantic-release | tee semantic-release-output.txt
+else
+  semantic-release -d | tee semantic-release-output.txt
+fi
 
 version=$(cat semantic-release-output.txt | grep "The next release version is" | rev | cut -d' ' -f1 | rev)
